Collect interfaces in single module
